# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Company Ship API",
        "description": "API quản lý vận tải - công ty, nhân sự, xe, chuyến đi, lương",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/departments": {
            "get": {
                "tags": [
                    "Departments"
                ],
                "summary": "Danh sách phòng ban",
                "operationId": "e8478e3a557995e1da564073521762f6",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Tìm theo code, name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "office_id",
                        "in": "query",
                        "description": "Lọc theo văn phòng",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sắp xếp",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Số bản ghi/trang",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                }
            },
            "post": {
                "tags": [
                    "Departments"
                ],
                "summary": "Tạo phòng ban mới",
                "operationId": "9a3a8ebf7326b72fd181c15603cdd21e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "name",
                                    "office_id"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "DEP001"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Phòng Nhân sự"
                                    },
                                    "office_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "parent_id": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo thành công"
                    },
                    "422": {
                        "description": "Validation lỗi"
                    }
                }
            }
        },
        "/api/departments/{id}": {
            "get": {
                "tags": [
                    "Departments"
                ],
                "summary": "Chi tiết phòng ban",
                "operationId": "ba8d49f7da851631d9985375effc9b2d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            },
            "put": {
                "tags": [
                    "Departments"
                ],
                "summary": "Cập nhật phòng ban",
                "operationId": "883e32516e81142e7441bad60554a993",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "office_id": {
                                        "type": "integer"
                                    },
                                    "parent_id": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    },
                    "422": {
                        "description": "Validation lỗi"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Departments"
                ],
                "summary": "Xóa phòng ban",
                "operationId": "4f695822af3d4a6db6f902305be6d0de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Dùng token trả về từ POST /api/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}